Mueve la lógica de ventas a su propia store
